+++ Plugins/Effect/Effect.cpp
@@ -4,6 +4,9 @@
 #include "API/Globals.hpp"
 #include "API/CExoString.hpp"
 #include "API/CGameEffect.hpp"
+#include "API/Functions.hpp"
+#include "API/CVirtualMachine.hpp"
+#include "API/CNWSObject.hpp"
 #include "Utils.hpp"
 #include "ViewPtr.hpp"
 
@@ -11,6 +14,7 @@
 #include <functional>
 
 using namespace NWNXLib;
+using namespace NWNXLib::API;
 
 static ViewPtr<Effect::Effect> g_plugin;
 
@@ -44,6 +48,8 @@ Effect::Effect(const Plugin::CreateParams& params)
 
     REGISTER(PackEffect);
     REGISTER(UnpackEffect);
+    REGISTER(SetEffectExpiredScript);
+    REGISTER(GetEffectExpiredData);
 
 #undef REGISTER
 
@@ -167,5 +173,60 @@ ArgumentStack Effect::UnpackEffect(ArgumentStack&& args)
     return stack;
 }
 
+ArgumentStack Effect::SetEffectExpiredScript(ArgumentStack&& args)
+{
+    static bool bOnEffectRemovedHook;
+
+    if (!bOnEffectRemovedHook)
+    {
+        GetServices()->m_hooks->RequestSharedHook<API::Functions::CNWSEffectListHandler__OnEffectRemoved, int32_t>(
+            +[](Services::Hooks::CallType type, CNWSEffectListHandler*, CNWSObject* pObject, CGameEffect* pEffect) -> void
+            {
+                if (type == Services::Hooks::CallType::BEFORE_ORIGINAL)
+                {
+                    CExoString &sScriptName = pEffect->m_sParamString[4];
+                    if (!sScriptName.IsEmpty())
+                    {
+                        g_plugin->m_effectExpiredData = std::string(pEffect->m_sParamString[5].CStr());
+
+                        LOG_DEBUG("Running script '%s' on object '%x' with data '%s'", sScriptName.CStr(), pObject->m_idSelf, g_plugin->m_effectExpiredData.c_str());
+
+                        ++g_plugin->m_effectExpiredDepth;
+                        Globals::VirtualMachine()->RunScript(&sScriptName, pObject->m_idSelf, 1);
+                        --g_plugin->m_effectExpiredDepth;
+                    }
+                }
+            });
+
+        bOnEffectRemovedHook = true;
+    }
+
+    ArgumentStack stack;
+
+    auto effect = Services::Events::ExtractArgument<API::CGameEffect*>(args);
+
+    // Script name
+    effect->m_sParamString[4] = Services::Events::ExtractArgument<std::string>(args).c_str();
+    // Data
+    effect->m_sParamString[5] = Services::Events::ExtractArgument<std::string>(args).c_str();
+
+    Services::Events::InsertArgument(stack, effect);
+
+    return stack;
+}
+
+ArgumentStack Effect::GetEffectExpiredData(ArgumentStack&&)
+{
+    if (g_plugin->m_effectExpiredDepth == 0)
+    {
+        throw std::runtime_error("Attempted to get effect expired data in an invalid context.");
+    }
+
+    ArgumentStack stack;
+
+    Services::Events::InsertArgument(stack, g_plugin->m_effectExpiredData);
+
+    return stack;
+}
 
 }