+++ Plugins/Effect/Effect.hpp
@@ -2,6 +2,7 @@
 
 #include "Plugin.hpp"
 #include "Services/Events/Events.hpp"
+#include "Services/Hooks/Hooks.hpp"
 #include "API/Types.hpp"
 
 using ArgumentStack = NWNXLib::Services::Events::ArgumentStack;
@@ -15,9 +16,13 @@ public:
     virtual ~Effect();
 
 private:
+    std::string m_effectExpiredData;
+    uint32_t m_effectExpiredDepth;
+
     ArgumentStack PackEffect(ArgumentStack&& args);
     ArgumentStack UnpackEffect(ArgumentStack&& args);
-
+    ArgumentStack SetEffectExpiredScript(ArgumentStack&& args);
+    ArgumentStack GetEffectExpiredData(ArgumentStack&& args);
 };
 
 }